chore: Fix compiler errors in lv_libpng

+++ src/libs/libpng/lv_libpng.c
@@ -91,7 +91,7 @@ static lv_result_t decoder_info(lv_image_decoder_t * decoder, const void * src, 
          * [20..23]: height
          */
         uint8_t buf[24];
-        uint32_t rn;
+        size_t rn;
         lv_fs_read(&f, buf, sizeof(buf), &rn);
         lv_fs_close(&f);
 
@@ -184,8 +184,8 @@ static uint8_t * alloc_file(const char * filename, uint32_t * size)
 {
     uint8_t * data = NULL;
     lv_fs_file_t f;
-    uint32_t data_size;
-    uint32_t rn;
+    size_t data_size;
+    size_t rn;
     lv_fs_res_t res;
 
     *size = 0;
